MISUV-9576:   * refactor It/test => FinancialDetailsHipConnectorISpec;
